maint: rename CollectLoop related code to improve readability (#1722)
## Which problem is this PR solving?

The current naming for the collector (manager) and collect_loop(worker)
doesn't reflect the architecture design. This makes naming metrics and
understanding code logic hard. We would like to rename the newly added
code to reflect the new architecture so that it's easier to understand
and work with

## Short description of the changes

Type & Struct Renames:

- CollectLoop → CollectorWorker
- NewCollectLoop() → NewCollectorWorker()

Field & Variable Renames in InMemCollector:

- collectLoops []*CollectLoop → workers 
  []*CollectorWorker
- collectLoopsWG → workersWG
- houseKeepingWG → monitorWG
- outgoingTraces → tracesToSend

Method Renames:

- houseKeeping() → monitor()
- getLoopForTrace() → getWorkerIDForTrace()

Constant Renames:

- collectorLoopHealthPrefix → collectWorkerHealthPrefix

File Renames:
- collect/collect_loop.go → collect/collector_worker.go