fix: collect_cache_entries tracks total cache size, new worker_cache_entries for tracking worker cache sizes (#1721)
## Which problem is this PR solving?

`collect_cache_entries` is supposed to be the total of trace cache size.
It is currently reporting per-worker trace cache size. This PR fixes
that.

However, I did realize that we already have `collector_cache_size`
reporting the same metric. I wonder if we can combine them or remove one
of them

## Short description of the changes

- move the pre-existing `collect_cache_entries` metric back to the
InMemCollector
  - track it in the monitor goroutine (formerly housekeeping)
- add a new `worker_cache_entries` metric to InMemCache
  - each worker has an instance of an InMemCache
  - track worker cache entries here