perf: do not extract sampling fields for OTLP endpoint (#1748)
## Which problem is this PR solving?

In large Refinery deployments, the majority of incoming spans are
forwarded to peer nodes for their final sampling decision. This means
the ingesting node does not need to extract or memoize the sampling key
fields during initial unmarshaling. However, the current implementation
still performs this work for every span, even when the local node will
not use the sampling key. This results in unnecessary allocations and
higher memory usage. By skipping sampling key memoization on ingest, we
can significantly reduce allocation overhead and improve memory
efficiency.

This optimization currently applies only to OTLP data, because OTLP is
accepted exclusively from external sources and not from peers. While we
ultimately want to extend this optimization to all incoming data,
achieving that will require substantial additional work. We plan to
address this more comprehensive implementation in the future.

## Short description of the changes

- create a `UnmarshalMsgpEventMetadataOnly` method
- use the new method in OTLP path