perf: don't make another copy of Event in Span struct (#1747)
## Which problem is this PR solving?

Because a Span object lives for the duration of the sampling process in
Refinery, it is safe to store a pointer to the original Event inside the
Span struct. Doing so avoids performing an unnecessary copy of the Event
when creating a Span, reducing allocation and memory overhead.

## Short description of the changes

Update the Span struct to store a pointer to the original Event instead
of copying the Event data