docs: update 3.0.0 RELNOTES with parent ID rule breaking change
`trace.parent_id` (or any field configured in ParentNames) is used to mark a span as "root" in the internal data structure now. It is not memoized as a metadata or key field available for rule matching. Rule matching the root span is why the `has-root-span` operator exists.

Document this breaking change for rules for customers to update their rules during 3.0 upgrade.